changed from popup to a whole page
